Refactor module 'game'.
Game now contains the state, and Game::run will modify the game struct.
Also, I made game processing functions not a methods of game but a free function that receives game as an argument. This feels more intuitive when using continuations, gives better type inference, and removes the need to pass the reference of whole game structure.